requirements.txt
matplotlib
Pillow
opencv-python
numpy
Flask
setuptools
pylint
pytest
pytest-cov
pytest-pylint
pytest-qt
flake8
pytest-xvfb
WTForms
Werkzeug
easygui
Flask-JSGlue
click
itsdangerous
jinja2 == 3.0
MarkupSafe
Flask-WTF
absl-py
ann_visualizer
argon2-cffi
argon2-cffi-bindings
astroid
asttokens
astunparse
atomicwrites
attrs
backcall
beautifulsoup4
black
bleach
cachetools
certifi
cffi
charset-normalizer
click
colorama
coverage
cycler
debugpy
decorator
defusedxml
dill
entrypoints
etils
executing
fastjsonschema
Flask
flatbuffers
fonttools
gast
google-auth
google-auth-oauthlib
google-pasta
googleapis-common-protos
graphviz
grpcio
h5py
#idna
importlib-resources
iniconfig
install
ipykernel
ipython
ipython-genutils
ipywidgets
isort
itsdangerous
jedi
joblib
jsonschema
jupyter-client
jupyter-core
jupyterlab-pygments
jupyterlab-widgets
kennard-stone
keras
Keras-Preprocessing
keras-tuner
kiwisolver
lazy-object-proxy
libclang
Markdown
MarkupSafe
matplotlib
matplotlib-inline
mccabe
mistune
modAL
mypy-extensions
nbclient
nbconvert
nbformat
nest-asyncio
notebook
oauthlib
opt-einsum
packaging
pandocfilters
parso
pathspec
#pickleshare
platformdirs
pluggy
prometheus-client
promise
prompt-toolkit
protobuf
psutil
pure-eval
py
pyasn1
pyasn1-modules
pycodestyle
pycparser
pydot
pyflakes
Pygments
pyparsing
pyrsistent
pytest-xvfb
python-dateutil
pytz
PyVirtualDisplay
pyzmq
requests
requests-oauthlib
rsa
scikeras
scikit-learn
scipy
#seaborn
Send2Trash
six
soupsieve
stack-data
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow
tensorflow-datasets
tensorflow-estimator
tensorflow-io-gcs-filesystem
tensorflow-metadata
termcolor
terminado
threadpoolctl
toml
tomli
tomlkit
tornado
tqdm
traitlets
typing_extensions
urllib3
wcwidth
webencodings
widgetsnbextension
wrapt
zipp
pandas
palettable
scipy
SimpleITK
visualkeras

